+++ plugins/matomo-backend/config.d.ts
@@ -1,0 +1,12 @@
+export interface Config {
+  matomo: {
+    /**
+     * @visibility backend
+     */
+    apiToken: string;
+    /**
+     * @visibility backend
+     */
+    apiUrl: string;
+  };
+}