+++ plugins/rh-rbac-common/src/index.ts
@@ -1,8 +1,0 @@
-/**
- * Common functionalities for the rh-rbac plugin.
- *
- * @packageDocumentation
- */
-
-export * from './types';
-export * from './permissions';